deletehashtable

+++ hash_tables/6-hash_table_delete.c
@@ -6,7 +6,7 @@
  */
 void hash_table_delete(hash_table_t *ht)
 {
-unsigned long int index = 0;
+unsigned int index = 0;
 hash_node_t *node1 = NULL;
 hash_node_t *node2 = NULL;
 if(!ht)
